+++ tiger-common/src/main/java/de/gematik/test/tiger/common/data/config/TigerServerTypeSerializer.java
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.databind.SerializerProvider;
 import de.gematik.test.tiger.testenvmgr.servers.TigerServerType;
 import java.io.IOException;
 
-public class TigerServerTypeSerlializer extends JsonSerializer<TigerServerType> {
+public class TigerServerTypeSerializer extends JsonSerializer<TigerServerType> {
     @Override
     public void serialize(TigerServerType tigerServerType, JsonGenerator jsonGenerator, SerializerProvider serializerProvider)
         throws IOException {