+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/servers/AbstractTigerServer.java
@@ -150,7 +150,7 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
 
     private void reloadConfiguration() {
         try {
-            this.configuration = TigerGlobalConfiguration.instantiateConfigurationBeanStrict(CfgServer.class,
+            this.configuration = TigerGlobalConfiguration.instantiateConfigurationBeanStrict(getConfigurationBeanClass(),
                     "tiger", "servers", getServerId())
                 .orElseThrow(
                     () -> new TigerEnvironmentStartupException("Could not reload configuration for server with id %s", getServerId()));
@@ -158,6 +158,10 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
         } catch (TigerConfigurationException e) {
             log.warn("Could not reload configuration for server {}", getServerId(), e);
         }
+    }
+
+    public Class<? extends CfgServer> getConfigurationBeanClass() {
+        return CfgServer.class;
     }
 
     private void loadPkiForProxy() {
@@ -348,7 +352,7 @@ public abstract class AbstractTigerServer implements TigerEnvUpdateSender {
         }
     }
 
-    void publishNewStatusUpdate(TigerServerStatusUpdate update) {
+    public void publishNewStatusUpdate(TigerServerStatusUpdate update) {
         update.setType(getServerTypeToken());
         tigerTestEnvMgr.publishStatusUpdateToListeners(TigerStatusUpdate.builder()
                         .serverUpdate(new LinkedHashMap<>(Map.of(serverId, update)))