+++ tiger-zion/src/main/java/de/gematik/test/tiger/zion/config/ZionConfiguration.java
@@ -14,4 +14,5 @@ public class ZionConfiguration {
     private Map<String, TigerMockResponse> mockResponses = new HashMap<>();
     private Map<String, String> mockResponseFiles = new HashMap<>();
     private ZionSpyConfiguration spy;
+    private int serverPort;
 }